Added check for valid objectid

+++ app/jobs/src/jobs.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -6,25 +7,27 @@ import {
   Param,
   ParseIntPipe,
   Post,
-  UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
 import { JobsService } from './jobs.service';
 import { CreateJobDto } from './dto/create-job.dto';
 import { Job } from './schemas/job.schema';
-import { GetByIdDto } from './dto/get-by-id.dto';
+import { isValidObjectId } from 'mongoose';
 
 @Controller('/api/jobs')
 export class JobsController {
   constructor(private readonly jobsService: JobsService) {}
 
   @Get()
-  getHello(): Promise<Job[]> {
+  getAllJobs(): Promise<Job[]> {
     return this.jobsService.getAllJobs();
   }
 
   @Get('/:id')
   getJobById(@Param('id') id: string): Promise<Job> {
+    //Make this in to a pipe later
+    if(!isValidObjectId(id)) throw new BadRequestException(`${id} is not a valid ObjectId`)
+
     return this.jobsService.getJobById(id);
   }
 
@@ -34,7 +37,10 @@ export class JobsController {
   }
 
   @Delete('/:id')
-  deleteJob(@Param('id', ParseIntPipe) id: number) {
+  deleteJob(@Param('id') id: string) {
+    //Make this in to a pipe later
+    if(!isValidObjectId(id)) throw new BadRequestException(`${id} is not a valid ObjectId`)
+
     return this.jobsService.deleteJob(id);
   }
 }
